requirements.txt
pdf2image
Pillow
requests
PyYAML
tqdm
retry